# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Real Estate Operations Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── LEADS ───────────────────────────────────────────────────────────────────
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/notes:
    get:
      operationId: getClientNotes
      tags: [clients]
      summary: Get notes for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createClientNote
      tags: [clients]
      summary: Add a note to a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /clients/{id}/activity:
    get:
      operationId: getClientActivity
      tags: [clients]
      summary: Get activity timeline for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"

  # ─── TASKS ───────────────────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Appointment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── WORKFLOWS (AUTOMATION) ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        role: { type: string }
        status: { type: string }
        avatarUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    AgentInput:
      type: object
      required: [name, email, role]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
        avatarUrl: { type: string }
    AgentUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
        status: { type: string }
        avatarUrl: { type: string }
...
    Client:
      type: object
      required: [id, firstName, lastName, email, status, createdAt]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        status: { type: string }
        budget: { type: ["number", "null"] }
        preferredPropertyType: { type: ["string", "null"] }
        assignedAgentId: { type: ["integer", "null"] }
        assignedAgentName: { type: ["string", "null"] }
        createdAt: { type: string }
    ClientInput:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string }
        budget: { type: number }
        preferredPropertyType: { type: string }
        assignedAgentId: { type: integer }
    ClientUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string }
        budget: { type: number }
        preferredPropertyType: { type: string }
        assignedAgentId: { type: ["integer", "null"] }

    # ─── NOTE ────────────────────────────────────────────────────────────────────
    Note:
      type: object
      required: [id, content, entityType, entityId, createdAt]
      properties:
        id: { type: integer }
        content: { type: string }
        entityType: { type: string }
        entityId: { type: integer }
        authorName: { type: ["string", "null"] }
        createdAt: { type: string }
    NoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        authorName: { type: string }

    # ─── TASK ────────────────────────────────────────────────────────────────────
...
    Appointment:
      type: object
      required: [id, title, type, startTime, endTime, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        status: { type: string }
        location: { type: ["string", "null"] }
        assignedAgentId: { type: ["integer", "null"] }
        assignedAgentName: { type: ["string", "null"] }
        leadId: { type: ["integer", "null"] }
        clientId: { type: ["integer", "null"] }
        propertyId: { type: ["integer", "null"] }
        createdAt: { type: string }
    AppointmentInput:
      type: object
      required: [title, type, startTime, endTime]
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        status: { type: string }
        location: { type: string }
        assignedAgentId: { type: integer }
        leadId: { type: integer }
        clientId: { type: integer }
        propertyId: { type: integer }
    AppointmentUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        status: { type: string }
        location: { type: string }
        assignedAgentId: { type: ["integer", "null"] }
        leadId: { type: ["integer", "null"] }
        clientId: { type: ["integer", "null"] }
        propertyId: { type: ["integer", "null"] }

    # ─── WORKFLOW ────────────────────────────────────────────────────────────────
...
    ActivityLog:
      type: object
      required: [id, action, entityType, entityId, description, createdAt]
      properties:
        id: { type: integer }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: integer }
        description: { type: string }
        performedBy: { type: ["string", "null"] }
        createdAt: { type: string }