# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real Estate Operations Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentUpdate {
  name?: string;
  email?: string;
  phone?: string;
  role?: string;
  status?: string;
  avatarUrl?: string;
}